+++ src/vscode_debug_console/vscode_debug_console.ts
@@ -1,8 +1,12 @@
 import * as vscode from 'vscode';
-import { logError, logInfo, showWarning } from '../utils/src/logger/logger';
+import { Icon_Info, logError, showWarning } from '../utils/src/logger/logger';
 import { getYAMLFileContent, readFileToText } from '../utils/src/vscode_utils/editor_utils';
 import { checkGitExtensionInYamlIfDart } from '../utils/src/language_utils/dart/pubspec/analyze_dart_git_dependency';
 import * as path from 'path';
+import { loadConfig } from './config_state';
+import { fileURLToPath, pathToFileURL } from 'url';
+import * as fs from 'fs';
+import { runTerminal } from '../utils/src/terminal_utils/terminal_utils';
 // --- 全域狀態管理 ---
 let dartToolPackage = new Map<String, any>()
 let projectLibAbsPath = new Map<String, String>()
@@ -11,11 +15,22 @@ let sessionToolConfigPath = ""
 let sessionProjectPackageName = ""
 // 檔案監聽器
 let pubspecWatcher: vscode.FileSystemWatcher | undefined;
-
 let workspaceRootPath = ""
+let silent = true
+const watchers: vscode.FileSystemWatcher[] = [];
+
+let allFiles: string[] = [];
 export class APP {
     public static flutterYaml: any | undefined = undefined;
     public static flutterPackageName: any | undefined = undefined;
+}
+
+
+function logInfo(msg: string = "", showOnVscode: boolean = true) {
+    console.log(`${Icon_Info} : ${msg}`);
+    if (showOnVscode && !silent) {
+        vscode.window.showInformationMessage(msg)
+    }
 }
 
 export async function registerDebugConsole(context: vscode.ExtensionContext) {
@@ -23,7 +38,7 @@ export async function registerDebugConsole(context: vscode.ExtensionContext) {
         workspaceRootPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
         logInfo(`Workspace root detected: ${workspaceRootPath}`);
     }
-
+    await updateFiles();
     // 監聽偵錯 session 的啟動
     context.subscriptions.push(vscode.debug.onDidStartDebugSession(async (session) => {
         const cwd = session.configuration.cwd;
@@ -46,11 +61,28 @@ export async function registerDebugConsole(context: vscode.ExtensionContext) {
 
     // 註冊 Debug Adapter Tracker，用於修改除錯控制台的輸出
     context.subscriptions.push(vscode.debug.registerDebugAdapterTrackerFactory('*', {
+
         createDebugAdapterTracker(session: vscode.DebugSession) {
             return {
                 async onDidSendMessage(message: any) {
                     // 只處理有 sessionPath 的情況
                     if (!sessionPath) return;
+                    // Normalize and rewrite file paths in build-time error messages
+                    if (message.type === 'event' && message.event === 'output') {
+                        const pattern = /([\w/.-]+):(\d+):(\d+)/;
+                        const match = message.body.output.match(pattern);
+                        let filePath = '';
+                        if (match) {
+                            filePath = match[1];
+                        }
+                        if (filePath.startsWith('lib/')) {
+                            let r = allFiles.filter((e) => e.includes(filePath))
+                            if (r.length > 0) {
+                                const relativePath = path.relative(workspaceRootPath, `${sessionPath}/${filePath}`);
+                                message.body.output = message.body.output.replace(filePath, `./${relativePath}`)
+                            }
+                        }
+                    }
                     // 核心邏輯：攔截並修改包含 'package:' 的輸出日誌
                     if (message.type === 'event' && message.event === 'output' && (message.body.category === 'stdout' || message.body.category === 'console')) {
                         // 检查消息内容中是否包含 'package:' 前缀
@@ -82,32 +114,67 @@ export async function registerDebugConsole(context: vscode.ExtensionContext) {
                             if (message.body.output.includes('package:') || message.body.output.includes('(packages/')) {
                                 findPackage = filePath.split('/')[0];
                             }
-                            let fullPath = ""
+                            let absPath = '';
+                            let fullPath = '';
+                            const config = loadConfig()
+                            silent = config.silent
+                            const relativePathMode = config.relativePathMode
+                            let sessionProjectLog = false
                             for (let key of dartToolPackage.keys()) {
                                 let value = dartToolPackage.get(key)
                                 let packages = value.packages
                                 let packageRef
-                                for (let p of packages) {
-                                    const resourceUri = vscode.window.activeTextEditor?.document.uri;
+                                for (let packageRef of packages) {
+                                    if (packageRef.name === findPackage) {
 
-                                    if (p.name === findPackage) {
-                                        const packageRef = p;
+                                        // Handle session project
                                         if (packageRef.rootUri === "../") {
-                                            let absPath = projectLibAbsPath.get(findPackage)
-                                            const relative = path.relative(workspaceRootPath, absPath as string);
-                                            if (relative === "") {
-                                                fullPath = "./lib"
-                                            } else {
-                                                fullPath = "./" + relative + "/" + "lib/"
+                                            let absPath = projectLibAbsPath.get(findPackage) as string
+                                            const relativePath = path.relative(workspaceRootPath, absPath);
+                                            const sessionOnRoot = relativePath == ""
+                                            sessionProjectLog = true
+                                            switch (relativePathMode) {
+                                                case 'session':
+                                                case 'workspace':
+                                                case 'always':
+                                                    fullPath = sessionOnRoot ? './lib/' : `./${relativePath}/lib/`;
+                                                    break;
+                                                case 'never':
+                                                default:
+                                                    fullPath = `${toFileUri(absPath)}/lib/`;
+                                            }
+
+                                        }
+                                        // Session relative dependency
+                                        else if (packageRef.rootUri.startsWith("../")) {
+                                            absPath = path.resolve(sessionPath, packageRef.rootUri.replace("../", ""));
+                                            sessionProjectLog = false;
+                                            const relativePath = path.relative(workspaceRootPath, absPath);
+
+                                            const isSubdirectory = absPath.includes(workspaceRootPath);
+                                            if (relativePathMode === 'never') {
+                                                fullPath = `${toFileUri(absPath)}/lib/`;
+                                            } else if (relativePathMode === 'workspace' && !isSubdirectory) {
+                                                fullPath = `${toFileUri(absPath)}/lib/`;
+                                            }
+                                            else {
+                                                fullPath = `./${relativePath}/lib/`;
                                             }
                                         }
-                                        // local package
-                                        else if (packageRef.rootUri.startsWith("../")) {
-                                            const absPath = path.resolve(sessionPath, packageRef.rootUri.replace("../", ""));
-                                            fullPath = absPath + "/lib"
-                                        }
                                         else {
-                                            fullPath = packageRef.rootUri + "/lib"
+                                            absPath = packageRef.rootUri;
+                                            const filePath = fileURLToPath(absPath);
+                                            const relativePath = path.relative(workspaceRootPath, filePath);
+                                            const isSubdirectory = filePath.startsWith(workspaceRootPath);
+                                            const isSessionSubdirectory = filePath.startsWith(sessionPath);
+                                            sessionProjectLog = false;
+                                            if (relativePathMode === 'always' ||
+                                                (relativePathMode === 'workspace' && isSubdirectory) ||
+                                                (relativePathMode === 'session' && isSessionSubdirectory)) {
+                                                fullPath = `./${relativePath}/lib/`;
+                                            } else {
+                                                fullPath = `${absPath}/lib/`;
+                                            }
                                         }
 
                                     }
@@ -118,10 +185,18 @@ export async function registerDebugConsole(context: vscode.ExtensionContext) {
                                 return
                             }
                             let newOutput = ''
+                            const isSdk = message.body.output.includes("packages/flutter") || message.body.output.includes("dart:core");
+                            if (config.showEmoji) {
+                                let emoji = sessionProjectLog ? config.emojiMap["session"] : config.emojiMap["pub"]
+                                if (isSdk) {
+                                    emoji = config.emojiMap["sdk"]
+                                }
+                                fullPath = `${emoji} │ ${fullPath}`
+                            }
                             if (message.body.output.includes(`packages/${findPackage}`)) {
-                                newOutput = message.body.output.replace(`packages/${findPackage}`, ` ${fullPath}`)
+                                newOutput = message.body.output.replace(`packages/${findPackage}/`, ` ${fullPath}`)
                             } else {
-                                newOutput = message.body.output.replace(`package:${findPackage}`, ` ${fullPath}`)
+                                newOutput = message.body.output.replace(`package:${findPackage}/`, ` ${fullPath}`)
                             }
                             newOutput = newOutput.replace(` ${lineNumber}:${columnNumber}`, `:${lineNumber}:${columnNumber}`)
                             message.body.output = newOutput
@@ -137,11 +212,39 @@ export async function registerDebugConsole(context: vscode.ExtensionContext) {
     }
     )
     );
+    // 监听文件系统事件
+    vscode.workspace.onDidRenameFiles(async event => {
+        console.log('Files renamed.');
+        // 更新文件列表
+        await updateFiles();
+        console.log('Updated files:', allFiles);
+    });
 
+    vscode.workspace.onDidCreateFiles(async event => {
+        console.log('Files renamed.');
+        // 更新文件列表
+        await updateFiles();
+        console.log('Updated files:', allFiles);
+    });
+
+
+    vscode.workspace.onDidDeleteFiles(async event => {
+        console.log('Files deleted.');
+        // 更新文件列表
+        await updateFiles();
+        console.log('Updated files:', allFiles);
+    });
 
 }
 
-
+async function updateFiles() {
+    allFiles = []
+    await vscode.workspace.findFiles('**/*.dart').then((files) => {
+        files.forEach((file) => {
+            allFiles.push(file.path)
+        });
+    });
+}
 
 let refreshTimeout: NodeJS.Timeout | undefined = undefined;
 
@@ -167,7 +270,23 @@ function scheduleRefresh(projectDirectory: string, context: vscode.ExtensionCont
 }
 
 
+export interface PackageConfig {
+    configVersion: number;
+    packages: DartPackage[];
+    generated: string;
+    generator: string;
+    generatorVersion: string;
+    flutterRoot?: string;
+    flutterVersion?: string;
+    pubCache?: string;
+}
 
+export interface DartPackage {
+    name: string;
+    rootUri: string;
+    packageUri: string;
+    languageVersion: string;
+}
 
 /**
  * [優化核心] 直接分析指定目錄的專案結構，而不是全域搜尋。
@@ -192,7 +311,7 @@ async function traverseProject(projectDirectory: string): Promise<boolean> {
         // 3. 檢查 package_config.json 是否存在
         await vscode.workspace.fs.stat(vscode.Uri.file(packageConfigPath));
         const packageConfigText = await readFileToText(packageConfigPath);
-        const packageConfigJson = JSON.parse(packageConfigText);
+        const packageConfig: PackageConfig = JSON.parse(packageConfigText);
 
         // 4. 清理並更新全域狀態
         dartToolPackage.clear();
@@ -200,9 +319,9 @@ async function traverseProject(projectDirectory: string): Promise<boolean> {
 
         sessionProjectPackageName = pubspecContent['name'];
         sessionToolConfigPath = packageConfigPath;
-        dartToolPackage.set(sessionProjectPackageName, packageConfigJson);
+        dartToolPackage.set(sessionProjectPackageName, packageConfig);
         projectLibAbsPath.set(sessionProjectPackageName, projectDirectory);
-
+        watchPackageConfigs(packageConfig)
         logInfo(`Successfully analyzed project: ${sessionProjectPackageName}`);
 
         // 附帶的檢查邏輯
@@ -264,4 +383,100 @@ function setupFileWatcher(projectDirectory: string, configPathToWatch: string, c
 
     // 將監聽器加入到 subscriptions 中，確保 extension 停用時會被自動清理
     context.subscriptions.push(pubspecWatcher);
+}
+
+export function toFileUri(absolutePath: string): string {
+    const normalized = path.resolve(absolutePath);
+    return pathToFileURL(normalized).toString();
+}
+
+
+function isLocalPackage(pkg: DartPackage): boolean {
+    return !pkg.rootUri.includes('.pub-cache') && !pkg.rootUri.includes('packages/flutter');
+}
+
+function disposeWatchers(watchers: vscode.FileSystemWatcher[]) {
+    watchers.forEach((watcher) => watcher.dispose());
+    watchers.length = 0;
+}
+
+export async function watchPackageConfigs(packageConfig: PackageConfig) {
+    disposeWatchers(watchers)
+    for (const pkg of packageConfig.packages) {
+        if (pkg.name === "utils") {
+            let a
+        }
+        if (!isLocalPackage(pkg)) continue;
+        let absPath = pkg.rootUri;
+        if (pkg.rootUri.startsWith("..")) {
+            absPath = path.resolve(sessionPath, pkg.rootUri.replace("../", ""));
+        }
+        let localPath: string;
+        try {
+            localPath = vscode.Uri.parse(absPath).fsPath;
+        } catch (e) {
+            console.warn(`Invalid URI: ${pkg.rootUri}`);
+            continue;
+        }
+
+        const configPath = path.join(localPath, '.dart_tool', 'package_config.json');
+
+        if (!fs.existsSync(configPath)) continue;
+
+        const watcher = vscode.workspace.createFileSystemWatcher(configPath);
+        logInfo(`creating watcher for package_config.json in ${pkg.name} at ${configPath}`);
+        const delayMs = 5000;
+
+        watcher.onDidChange(() => {
+            throttleRun(pkg.name, delayMs, async () => {
+                console.log(`[FlutterLogger] package_config.json changed in ${pkg.name}, running pub get`);
+                await runFlutterPubGet();
+            });
+        });
+
+        watcher.onDidCreate(() => {
+            throttleRun(pkg.name, delayMs, async () => {
+                console.log(`[FlutterLogger] package_config.json created in ${pkg.name}, running pub get`);
+                await runFlutterPubGet();
+            });
+        });
+
+        watchers.push(watcher);
+    }
+
+    return watchers;
+}
+
+async function runFlutterPubGet() {
+    let terminal = findTerminalAndActivate("LazyJ Logger", sessionPath)
+    terminal.sendText('flutter pub get');
+    terminal.show();
+}
+
+function findTerminalAndActivate(name: string, pwd: string): vscode.Terminal {
+    const terminal = vscode.window.terminals.find(t => t.name == name);
+    if (terminal) {
+        terminal.show();
+        return terminal;
+    }
+    else {
+        const newTerminal = vscode.window.createTerminal(name);
+        newTerminal.sendText(`cd ${sessionPath}`)
+        newTerminal.show();
+        return newTerminal;
+    }
+}
+
+
+const throttleMap = new Map<string, NodeJS.Timeout>();
+
+function throttleRun(key: string, delayMs: number, fn: () => void) {
+    if (throttleMap.has(key)) return;
+
+    fn();
+    const timeout = setTimeout(() => {
+        throttleMap.delete(key);
+    }, delayMs);
+
+    throttleMap.set(key, timeout);
 }